+++ app/main.py
@@ -5,6 +5,7 @@ from app.db.session import get_db
 from app.core.config import settings
 from app.api.v1.routers import users, auth, locations, recommendations, user_events, bookmarks
 from app.api.v1.routers import itinerary
+from app.api.v1.routers import itinerary_step
 
 app = FastAPI()  
 
@@ -16,6 +17,7 @@ app.include_router(recommendations.router)
 app.include_router(users.router)
 app.include_router(user_events.router)
 app.include_router(itinerary.router)
+app.include_router(itinerary_step.router)
 
 
 @app.get("/")
